requirements.txt
# Testing dependencies
pynose
pinocchio
factory-boy

# Code Coverage
coverage

requests
